# rest/docs/swagger.yaml
openapi: 3.0.0
info:
    title: COVID and MPOX Data API
    description: API for accessing and managing COVID-19 and MPOX epidemic data
    version: 1.0.0
    contact:
        email: api-support@yourdomain.com
servers:
    - url: https://api.yourdomain.com/api
      description: Production server
    - url: https://staging-api.yourdomain.com/api
      description: Staging server
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: token
            description: Enter the API token as a bearer token
    schemas:
        CovidData:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                    description: Unique identifier
                date:
                    type: string
                    format: date
                    description: Date of the record
                country:
                    type: string
                    description: Country name
                total_cases:
                    type: integer
                    description: Total confirmed cases
                new_cases:
                    type: integer
                    description: New cases on this date
                active_cases:
                    type: integer
                    description: Currently active cases
                total_deaths:
                    type: integer
                    description: Total deaths
                new_deaths:
                    type: integer
                    description: New deaths on this date
                total_recovered:
                    type: integer
                    description: Total recovered cases
                daily_recovered:
                    type: integer
                    description: New recoveries on this date
            required:
                - date
                - country
                - total_cases
                - new_cases
                - active_cases
                - total_deaths
                - new_deaths
                - total_recovered
                - daily_recovered
...
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
                    description: Error message
            required:
                - error
        CovidTotals:
            type: object
            properties:
                total_cases:
                    type: string
                    description: Total confirmed cases worldwide
                total_deaths:
                    type: string
                    description: Total deaths worldwide
                total_recovered:
                    type: string
                    description: Total recovered cases worldwide
            required:
                - total_cases
                - total_deaths
                - total_recovered
...
paths:
    /covid/public/latest:
        get:
            summary: Get latest COVID data
            description: Returns the 10 most recent COVID data entries
            tags:
                - Public COVID Data
            responses:
                "200":
                    description: A list of the latest COVID data entries
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/CovidData"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /covid/public/country/{country}:
        get:
            summary: Get COVID data for a specific country
            description: Returns the last 30 days of COVID data for a specific country
            tags:
                - Public COVID Data
            parameters:
                - name: country
                  in: path
                  required: true
                  description: Country name
                  schema:
                      type: string
            responses:
                "200":
                    description: COVID data for the specified country
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/CovidData"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /covid/public/totals:
        get:
            summary: Get global COVID totals
            description: Returns global COVID totals from the most recent data
            tags:
                - Public COVID Data
            responses:
                "200":
                    description: Global COVID totals
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CovidTotals"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /mpox/public/summary:
        get:
            summary: Get MPOX summary by country
            description: Returns MPOX summary data grouped by country
            tags:
                - Public MPOX Data
            responses:
                "200":
                    description: MPOX summary by country
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        country:
                                            type: string
                                        latest_cases:
                                            type: integer
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...